+++ src/paco/cftemplates/sns.py
@@ -1,20 +1,21 @@
 """
-CloudFormation template for SNS Topics
+Template for SNS Topics and Subscriptions
 """
 
-import awacs.sns
-import troposphere
-import troposphere.sns
 from paco.cftemplates.cftemplates import StackTemplate
 from paco.models import references
 from paco.models import schemas
 from paco.models.locations import get_parent_by_interface
 from awacs.aws import Allow, Policy, Statement, Principal, Condition, StringEquals
+import awacs.sns
+import json
+import troposphere
+import troposphere.sns
 
 
-class SNSTopics(StackTemplate):
+class SNS(StackTemplate):
     """
-    CloudFormation template for SNS Topics
+SNS Topics and Subscriptions
     """
     def __init__(
         self,
@@ -47,12 +48,12 @@ class SNSTopics(StackTemplate):
         )
 
         if grp_id == None:
-            self.set_aws_name('SNSTopics', self.resource_group_name, self.resource_name)
+            self.set_aws_name('SNS', self.resource_group_name, self.resource_name)
         else:
-            self.set_aws_name('SNSTopics', grp_id)
+            self.set_aws_name('SNS', grp_id)
 
         # Troposphere Template Initialization
-        self.init_template('SNS Topics')
+        self.init_template('SNS Topics and Subscriptions')
         template = self.template
 
         # Topic Resources and Outputs
@@ -68,26 +69,7 @@ class SNSTopics(StackTemplate):
             if topic.display_name:
                 cfn_export_dict['DisplayName'] = topic.display_name
 
-            # Subscriptions
-            if len(topic.subscriptions) > 0:
-                cfn_export_dict['Subscription'] = []
-            for subscription in topic.subscriptions:
-                sub_dict = {}
-                if references.is_ref(subscription.endpoint):
-                    param_name = 'Endpoint{}'.format(topic_logical_id)
-                    parameter = self.create_cfn_parameter(
-                        param_type = 'String',
-                        name = param_name,
-                        description = 'SNSTopic Endpoint value',
-                        value = subscription.endpoint,
-                    )
-                    endpoint = parameter
-                else:
-                    endpoint = subscription.endpoint
-                sub_dict['Endpoint'] = endpoint
-                sub_dict['Protocol'] = subscription.protocol
-                cfn_export_dict['Subscription'].append(sub_dict)
-
+            # Topic Resource
             topic_resource = troposphere.sns.Topic.from_dict(
                 'Topic' + topic_logical_id,
                 cfn_export_dict
@@ -97,6 +79,35 @@ class SNSTopics(StackTemplate):
             topic.topic_resource = topic_resource
             template.add_resource(topic_resource)
 
+            # Subscriptions
+            idx = 0
+            for subscription in topic.subscriptions:
+                sub_dict = {
+                    'TopicArn': troposphere.Ref(topic_resource)
+                }
+                if references.is_ref(subscription.endpoint):
+                    param_name = f'Endpoint{topic_logical_id}{idx}'
+                    parameter = self.create_cfn_parameter(
+                        param_type = 'String',
+                        name = param_name,
+                        description = 'Subscription Endpoint',
+                        value = subscription.endpoint,
+                    )
+                    endpoint = parameter
+                else:
+                    endpoint = subscription.endpoint
+                sub_dict['Endpoint'] = endpoint
+                sub_dict['Protocol'] = subscription.protocol
+                if subscription.filter_policy:
+                    sub_dict['FilterPolicy'] = json.loads(subscription.filter_policy)
+                subscription_logical_id = f"Subscription{topic_logical_id}{idx}"
+                sub_resource = troposphere.sns.SubscriptionResource.from_dict(
+                    subscription_logical_id,
+                    sub_dict
+                )
+                template.add_resource(sub_resource)
+                idx += 1
+
             # Topic Outputs
             if grp_id == None:
                 output_ref = stack.resource.paco_ref_parts
@@ -105,12 +116,12 @@ class SNSTopics(StackTemplate):
             self.create_output(
                 title='SNSTopicArn' + topic_logical_id,
                 value=troposphere.Ref(topic_resource),
-                ref=output_ref + '.arn'
+                ref=f'{output_ref}.arn'
             )
             self.create_output(
                 title='SNSTopicName' + topic_logical_id,
                 value=troposphere.GetAtt(topic_resource, "TopicName"),
-                ref=output_ref + '.name',
+                ref=f'{output_ref}.name',
             )
 
         # Cross-account access policy